+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.5.0](https://github.com/unzerdev/php-sdk/compare/1.1.4.2..1.1.5.0)
+### Added
+*   Add Support for basket `v2` resource.
+### Changed
+*   Add support for payment state `create` which can occur when using payment pages.
+*   Examples:
+    *   Use `v2/basket` resource for secured payment methods and payment pages.
+    *   Remove broken image-URLs of payment page examples.
+*   Several minor improvements.
+
 ## [1.1.4.2](https://github.com/unzerdev/php-sdk/compare/1.1.4.1..1.1.4.2)
 ### Added
 *   Enable PHP 8.1 compatibility.